+++ src/Service/PaymentMethodService.php
@@ -35,6 +35,7 @@ use Kiener\MolliePayments\Handler\Method\Przelewy24Payment;
 use Kiener\MolliePayments\Handler\Method\RivertyPayment;
 use Kiener\MolliePayments\Handler\Method\SatispayPayment;
 use Kiener\MolliePayments\Handler\Method\SofortPayment;
+use Kiener\MolliePayments\Handler\Method\SwishPayment;
 use Kiener\MolliePayments\Handler\Method\TrustlyPayment;
 use Kiener\MolliePayments\Handler\Method\TwintPayment;
 use Kiener\MolliePayments\Handler\Method\VoucherPayment;
@@ -465,6 +466,7 @@ class PaymentMethodService
             PayByBankPayment::class,
             MbWayPayment::class,
             MultibancoPayment::class,
+            SwishPayment::class,
             // IngHomePayPayment::class, // not allowed anymore
             // DirectDebitPayment::class, // only allowed when updating subsriptions, aka => not allowed anymore
         ];