+++ tests/Cypress/cypress/e2e/storefront/payment-methods/swish.cy.js
@@ -1,0 +1,41 @@
+import Devices from "Services/utils/Devices";
+import Session from "Services/utils/Session"
+// ------------------------------------------------------
+import PaymentAction from "Actions/storefront/checkout/PaymentAction";
+import DummyBasketScenario from "Scenarios/DummyBasketScenario";
+import CheckoutAction from "Actions/storefront/checkout/CheckoutAction";
+// ------------------------------------------------------
+
+
+const devices = new Devices();
+const session = new Session();
+
+const paymentAction = new PaymentAction();
+const scenarioDummyBasket = new DummyBasketScenario(1);
+const device = devices.getFirstDevice();
+
+
+describe('Swish', () => {
+
+    context(devices.getDescription(device), () => {
+
+        before(function () {
+            devices.setDevice(device);
+        })
+
+        beforeEach(() => {
+            session.resetBrowserSession();
+            devices.setDevice(device);
+        });
+
+        it('Swish is existing in checkout', () => {
+
+            scenarioDummyBasket.execute();
+
+            paymentAction.switchPaymentMethod('swish');
+        })
+
+    })
+
+})
+